Id check

+++ online_journal.py
@@ -1,4 +1,5 @@
-from online_jpournal_db import OnlineJournalDB
+from online_journal_db import OnlineJournalDB
+
 
 def onlineJournal():
     db = OnlineJournalDB()
@@ -13,33 +14,50 @@ def onlineJournal():
         try:
             choice = input().lower()
             if(choice == 'insert'):
-                #insert
-                jid = int(input("Please Enter Journal Id : "))
-                jt = input("Please Enter Journal Title (Max 200 ch.) : ")
-                jw = input("Please Enter Writer Name (Max 200 ch.) : ")
-                jc = input("Please Enter Writer/Publisher contact Number : ")
-                jp = input("Please Enter Journal Published Date (DD-MM-YYYY) : ")
-                j = input("Please Start Writing Journal here (Max 1500 ch. ) : ")
-                db.insert_journal(jid,jt, jw, jc, jp, j)
+                #checking ID
+                id = int(input("Please Enter Journal Id : "))
+                if db.fetch_id(id) == 1:
+                    print("ID already Exists!..")
+                    continue
+                else:
+                    #insert
+                    jt = input("Please Enter Journal Title (Max 200 ch.) : ")
+                    jw = input("Please Enter Writer Name (Max 200 ch.) : ")
+                    jc = input("Please Enter Writer/Publisher contact Number : ")
+                    jp = input("Please Enter Journal Published Date (DD-MM-YYYY) : ")
+                    j = input("Please Start Writing Journal here (Max 1500 ch. ) : ")
+                    db.insert_journal(jid,jt, jw, jc, jp, j)
 
             elif(choice == 'fetch'):
                 #Fetch All Details
                 db.fetch_all()
 
             elif(choice == 'delete'):
-                #Delete Journal
-                jid = int(input("Enter Journal ID to delete : "))
-                db.delete_journal(jid)
+                #checking ID
+                id = int(input("Please Enter Journal Id : "))
+                if db.fetch_id(id) == 0:
+                    print("ID does not Exist!..")
+                    continue
+                else:
+                    #Delete Journal
+                    jid = int(input("Enter Journal ID to delete : "))
+                    db.delete_journal(jid)
 
             elif(choice == 'update'):
-                #update Journal
-                jid = int(input("Enter Journal ID to Update Details : "))
-                jt = input("Please Enter New Journal Title (Max 200 ch.) : ")
-                jw = input("Please Enter New Writer Name (Max 200 ch.) : ")
-                jc = input("Please Enter New Writer/Publisher contact Number : ")
-                jp = input("Please Enter New Journal Published Date (DD-MM-YYYY) : ")
-                db.update_journal(jid,jt, jw, jc, jp)
-            
+                #checking ID
+                id = int(input("Please Enter Journal Id : "))
+                if db.fetch_id(id) == 0:
+                    print("ID does not Exist!..")
+                    continue
+                else:
+                    #update Journal
+                    jid = int(input("Enter Journal ID to Update Details : "))
+                    jt = input("Please Enter New Journal Title (Max 200 ch.) : ")
+                    jw = input("Please Enter New Writer Name (Max 200 ch.) : ")
+                    jc = input("Please Enter New Writer/Publisher contact Number : ")
+                    jp = input("Please Enter New Journal Published Date (DD-MM-YYYY) : ")
+                    db.update_journal(jid,jt, jw, jc, jp)
+                
             elif(choice == 'q'):
                 print("See You again!..")
                 break
